+++ climate-data-ca/template/download/variable.php
@@ -35,6 +35,12 @@
         </div>
 
         <p class="form-help col-10 col-sm-2 offset-1"><a href="<?php echo get_permalink(filtered_ID_by_path('about', $GLOBALS['vars']['current_lang'])); ?>#about-datasets"><?php _e('Learn more about datasets', 'cdc'); ?></a></p>
+        <div class="form-select col-4 offset-4 mt-3">
+            <div id="average-btn-group" class="btn-group btn-group-toggle w-100" data-toggle="buttons">
+                <label id="average-label-allyears" class="btn btn-outline-primary"> <input type="radio" name="download-dataset-type" id="download-allyears" autocomplete="off" value="allyears"><?php _e('All years', 'cdc'); ?></label>
+                <label id="average-label-30yaverage" class="btn btn-outline-primary active"> <input type="radio" name="download-dataset-type" id="download-30yaverage" autocomplete="off" value="30ygraph" checked><?php _e('30 year averages', 'cdc'); ?></label>
+            </div>
+        </div>
     </div>
 
     <div id="selection-dataset" class="form-layout-row row align-items-center">
@@ -43,7 +49,7 @@
         </p>
 
         <div class="form-select col-10 offset-1 col-sm-4 offset-sm-0">
-            <div id="format-btn-group" class="btn-group btn-group-toggle w-100" data-toggle="buttons">
+            <div id="dataset-btn-group" class="btn-group btn-group-toggle w-100" data-toggle="buttons">
                 <label id="format-label-cmip5" class="btn btn-outline-primary" style="border-top-left-radius: 25px;border-bottom-left-radius: 25px;padding: 13px;"> <input type="radio" name="download-dataset" id="download-cmip5" autocomplete="off" value="cmip5">CMIP 5</label>
                 <label id="format-label-cmip6" class="btn btn-outline-primary active" style="border-top-right-radius: 25px;border-bottom-right-radius: 25px;padding: 13px;"> <input type="radio" name="download-dataset" id="download-cmip6" autocomplete="off" value="cmip6" checked>CMIP 6</label>
             </div>